Switch to new ability input bindings. Allow abilities to be replaced. Cut MAX_ABILITIES down to 2. Switch from per-person respawn pool to per-team ticket system for Assault. Add grace period after launching where drones can still receive damage from direct drone attacks. Add 30 second delay between last turret death and core becoming vulnerable.

+++ src/load.cpp
@@ -303,9 +303,9 @@ void Loader::settings_load(const Array<DisplayMode>& modes, const DisplayMode& c
 
 		bindings->invert_y = Json::get_s32(gamepad, "invert_y", 0);
 		bindings->zoom_toggle = Json::get_s32(gamepad, "zoom_toggle", 0);
-		bindings->sensitivity_gamepad = u8(Json::get_s32(gamepad, "sensitivity_gamepad", 100));
+		bindings->sensitivity_gamepad = u16(Json::get_s32(gamepad, "sensitivity_gamepad", 100));
 		if (i == 0)
-			bindings->sensitivity_mouse = u8(Json::get_s32(gamepad, "sensitivity_mouse", 100));
+			bindings->sensitivity_mouse = u16(Json::get_s32(gamepad, "sensitivity_mouse", 100));
 		bindings->rumble = b8(Json::get_s32(gamepad, "rumble", 1));
 		gamepad = gamepad ? gamepad->next : nullptr;
 	}
